requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
black==24.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
dataclasses-json==0.6.4
distro==1.9.0
flake8==7.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
marshmallow==3.20.2
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.11.0
packaging==23.2
pathspec==0.11.1
platformdirs==3.5.1
pycodestyle==2.11.1
pydantic==2.6.0
pydantic_core==2.16.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
